requirements.txt
requests==2.28.1
google-cloud-firestore==2.7.1